code/requirements.txt
streamlit
numpy
h5py
torch
torchvision
opencv-python-headless
Pillow
scikit-image
plotly
google-generativeai
python-dotenv
nibabel
tensorflow